requirements.txt
absl-py==1.3.0
alabaster==0.7.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
conda-pack==0.7.0
cryptography==37.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dgl-cu113==0.9.1
dglgo==0.0.2
docopt==0.6.2
docutils==0.19
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.37.4
future==0.18.2
gdown==4.5.1
gensim==4.2.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.0
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
importlib-resources==5.9.0
ipykernel==6.16.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
littleutils==0.2.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
memory-profiler==0.60.0
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.6.3
nlpaug==1.1.11
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.5
numpydoc==1.4.0
oauthlib==3.2.2
ogb==1.3.4
outdated==0.2.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
rdkit-pypi==2022.3.5
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.97
six==1.16.0
smart-open==6.2.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
torch==1.12.0
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchaudio==0.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.22.2
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.11
vocab==0.0.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yarg==0.1.9
zipp==3.10.0